requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
soupsieve==2.5
urllib3==2.1.0
